backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.5",
    "posthog>=6.7.10",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
